Sonidos en Pong, y nuevas funcionalidades
