+++ .env
@@ -1,1 +1,1 @@
-FULL_SCREEN=true
+FULL_SCREEN=false
