+++ dataCircus.csv
@@ -1,5 +1,3 @@
-Daniel,5280,11
-Fabricio,8065,13
-AXL,5840,51
-Maria,5305,11
-Roberto,6190,64
+Gustavo,5330,11
+Fabricio,5275,10
+Maria,6040,33
