+++ src/circusOverlays/GameoverOverlay.java
@@ -72,14 +72,14 @@ public class GameoverOverlay {
                 (FRAME_HEIGHT - OverlayConstants.BG_LOSE_OVERLAY_HEIGHT) / 2, OverlayConstants.BG_LOSE_OVERLAY_WIDTH,
                 OverlayConstants.BG_LOSE_OVERLAY_HEIGHT, null);
 
-        g.setFont(new Font("Arial", Font.BOLD, (int) (56 * Constants.SCALE)));
+        g.setFont(new Font("Roboto", Font.BOLD, (int) (56 * Constants.SCALE)));
         g.setColor(new Color(14, 165, 233));
 
         metrics = g.getFontMetrics();
         g.drawString("Game Over", FRAME_WIDTH / 2 - metrics.stringWidth("Game Over") / 2,
                 (FRAME_HEIGHT - OverlayConstants.BG_LOSE_OVERLAY_HEIGHT) / 2 + (int) (100 * Constants.SCALE));
 
-        g.setFont(new Font("Arial", Font.BOLD, (int) (32 * Constants.SCALE)));
+        g.setFont(new Font("Roboto", Font.BOLD, (int) (32 * Constants.SCALE)));
         metrics = g.getFontMetrics();
         g.drawString("Score: " + CircusPlaying.getScore(),
                 FRAME_WIDTH / 2 - metrics.stringWidth("Score: " + CircusPlaying.getScore()) / 2,
