+++ src/pongObjects/PongPlayer.java
@@ -6,14 +6,20 @@ import static utils.Constants.FrameConstants.FRAME_WIDTH;
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.Graphics;
+
+import audio.AudioPlayer;
+import states.PongPlaying;
 import utils.Constants.PongConstants;
 
 public class PongPlayer {
+    private PongPlaying pongPlaying;
+
     private int posX, posY;
     private int goals = 0, dir = 0;
     private boolean player_left;
 
-    public PongPlayer(Boolean pleft) {
+    public PongPlayer(Boolean pleft, PongPlaying pongPlaying) {
+        this.pongPlaying = pongPlaying;
         this.player_left = pleft;
 
         if (pleft) {
@@ -21,13 +27,16 @@ public class PongPlayer {
         } else {
             posX = PongConstants.PLAYER_2_START;
         }
-
         posY = PongConstants.PLAYER_START;
-
     }
 
     public void set_goal() {
         this.goals++;
+        pongPlaying.getGame().getAudioPlayer().playSounds(AudioPlayer.GOAL);
+
+        pongPlaying.setPause(true);
+        if (goals >= 5)
+            pongPlaying.setWinner(true);
     }
 
     public void draw(Graphics g) {
@@ -68,7 +77,7 @@ public class PongPlayer {
             }
         } else {
             if (ballPosX + PongConstants.BALL_SIZE >= posX && ballPosY + PongConstants.BALL_SIZE >= posY
-                    && ballPosY <= posY + FRAME_HEIGHT) {
+                    && ballPosY <= posY + PongConstants.PLAYER_HEIGHT) {
                 ball.change_direction(false);
             }
         }
