+++ src/utils/CSVFile.java
@@ -9,21 +9,24 @@ import java.util.ArrayList;
 
 public class CSVFile {
     private static String fileNameCircus = "dataCircus.csv";
+    private static String fileNamePong = "dataPong.csv";
 
     public CSVFile() {
-        createCSV();
+        createBothCSV();
     }
 
-    private void createCSV() {
+    private void createBothCSV() {
         File file = new File(fileNameCircus);
-        if (file.exists())
+        File file2 = new File(fileNamePong);
+        if (file.exists() && file2.exists())
             return;
 
         // * Crea el CSV (No le agrego un header xq da problemas a la hora de sort)
-        try (FileWriter writer = new FileWriter(fileNameCircus);) {
+        try (FileWriter writer = new FileWriter(fileNameCircus); FileWriter writer2 = new FileWriter(fileNamePong);) {
         } catch (IOException e) {
             e.printStackTrace();
         }
+
     }
 
     public static void writeLine(FileWriter writer, String[] row) throws IOException {
